requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
tavily-python>=0.5.0
streamlit>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
